pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.4",
    "django-rest-swagger>=2.2.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
]
